pyproject.toml
[tool.poetry]
name = "t2-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
streamlit = "^1.32.2"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
tiktoken = "^0.6.0"
langchain = "^0.1.13"
langchain-community = "^0.0.29"
langchain-experimental = "^0.0.55"
langchain-openai = "^0.1.1"
langchain-google-genai = "^0.0.11"
langchainhub = "^0.1.15"
ipython = "^8.22.2"
ipykernel = "^6.29.3"
pdf2image = "^1.17.0"
streamlit-chat = "^0.1.1"
unstructured = "^0.12.6"
chromadb = "^0.4.24"
jupyterlab = "^4.1.5"
pillow = "^10.2.0"
wikipedia = "^1.4.0"
duckduckgo-search = "^5.1.0"
pypdf = "^4.1.0"
docx2txt = "^0.8"
openai = "^1.14.2"
pyyaml = "^6.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
